break parser into pieces
